requirement.txt
fastapi
uvicorn
httpx
google-genai
Pillow
python-dotenv